# openapi.yaml
openapi: 3.1.0
info:
  title: Lex-Advisor RAG Service
  version: 0.1.0
paths:
...
  /v1/ingest:
    post:
      tags:
      - Ingest
      - Ingest
      summary: Ingest Document
      operationId: ingest_document_v1_ingest_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Request-ID
        in: header
        required: true
        schema:
          type: string
          title: X-Request-Id
      - name: X-Tenant-ID
        in: header
        required: true
        schema:
          type: string
          title: X-Tenant-Id
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Bad Request - Format invalid
        '401':
          description: Unauthorized - Invalid API Key
        '403':
          description: Forbidden - Acces interzis
        '404':
          description: "Not Found - Resursa nu a fost g\u0103sit\u0103"
        '409':
          description: Duplicate Job - Conflict de Idempotency-Key
        '429':
          description: Rate Limit Exceeded - Prea multe cereri
        '500':
          description: Internal Server Error
        '503':
          description: "Service Unavailable - Dependen\u021Be indisponibile"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/ingest/{job_id}:
    get:
      tags:
      - Ingest
      - Ingest
      summary: Get Job Status
      operationId: get_job_status_v1_ingest__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request - Format invalid
        '401':
          description: Unauthorized - Invalid API Key
        '403':
          description: Forbidden - Acces interzis
        '404':
          description: "Not Found - Resursa nu a fost g\u0103sit\u0103"
        '409':
          description: Duplicate Job - Conflict de Idempotency-Key
        '429':
          description: Rate Limit Exceeded - Prea multe cereri
        '500':
          description: Internal Server Error
        '503':
          description: "Service Unavailable - Dependen\u021Be indisponibile"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/namespaces/{namespace_id}/sources/{source_id}:
    delete:
      tags:
      - Namespaces
      summary: Delete Single Source
      operationId: delete_single_source_v1_namespaces__namespace_id__sources__source_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          title: Source Id
      responses:
        '204':
          description: Successful Response
        '400':
          description: Bad Request - Format invalid
        '401':
          description: Unauthorized - Invalid API Key
        '403':
          description: Forbidden - Acces interzis
        '404':
          description: "Not Found - Resursa nu a fost g\u0103sit\u0103"
        '409':
          description: Duplicate Job - Conflict de Idempotency-Key
        '429':
          description: Rate Limit Exceeded - Prea multe cereri
        '500':
          description: Internal Server Error
        '503':
          description: "Service Unavailable - Dependen\u021Be indisponibile"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/namespaces/{namespace_id}:
    delete:
      tags:
      - Namespaces
      summary: Delete Namespace
      operationId: delete_namespace_v1_namespaces__namespace_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request - Format invalid
        '401':
          description: Unauthorized - Invalid API Key
        '403':
          description: Forbidden - Acces interzis
        '404':
          description: "Not Found - Resursa nu a fost g\u0103sit\u0103"
        '409':
          description: Duplicate Job - Conflict de Idempotency-Key
        '429':
          description: Rate Limit Exceeded - Prea multe cereri
        '500':
          description: Internal Server Error
        '503':
          description: "Service Unavailable - Dependen\u021Be indisponibile"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/namespaces/{namespace_id}/stats:
    get:
      tags:
      - Namespaces
      summary: Get Namespace Stats
      operationId: get_namespace_stats_v1_namespaces__namespace_id__stats_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: path
        required: true
        schema:
          type: string
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request - Format invalid
        '401':
          description: Unauthorized - Invalid API Key
        '403':
          description: Forbidden - Acces interzis
        '404':
          description: "Not Found - Resursa nu a fost g\u0103sit\u0103"
        '409':
          description: Duplicate Job - Conflict de Idempotency-Key
        '429':
          description: Rate Limit Exceeded - Prea multe cereri
        '500':
          description: Internal Server Error
        '503':
          description: "Service Unavailable - Dependen\u021Be indisponibile"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          title: Status
        submitted_at:
          type: string
          title: Submitted At
        estimated_completion_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Estimated Completion At
      type: object
      required:
      - job_id
      - status
      - submitted_at
      title: JobResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer